requirements.txt
python-telegram-bot==22.5
google-genai==1.46.0
requests==2.32.5
python-dotenv==1.2.1
SQLAlchemy==2.0.44
aiosqlite>=0.19.0

# Webhook deployment dependencies
# FastAPI provides an ASGI framework to expose a webhook endpoint when running
# on Vercel or Render. Uvicorn is an ASGI server used for running FastAPI apps.
fastapi==0.120.0
uvicorn==0.38.0